version4

+++ lescript.sh
@@ -1,24 +1,26 @@
 #!/bin/bash
 
 filelock=/tmp/out/*.lock
+exec  &> /tmp/out/fileerror.log
 
-if test -f $filelock 
+
+if test -f $filelock
 then
-	exit 2>>tmp/out/fileerror.log
+	exit >> /tmp/out/fileerror.log
 else
 	touch /tmp/out/file.lock
 	for files in /tmp/in/*
 	do
-		if test -e -f $files
+		if test -e $files && test -f $files
 		then
 			ls $files
 			gzip -r -f $files
 			mv $files.gz  /tmp/out/
 		else
 			rm /tmp/out/file.lock
-			echo $? 2>>tmp/out/fileerror.log
+			echo $? >> /tmp/out/fileerror.log
 		fi
 	done
 	rm /tmp/out/file.lock
-	echo "le fichier $files a été déplacé dans /tmp/out" 2>> /tmp/out/fileerror.log
+	echo "le fichier $files a été déplacé dans /tmp/out" >> /tmp/out/fileerror.log
 fi
